Version 1.0

+++ app.py
@@ -5,9 +5,11 @@ import sys
 import logging
 import time
 import uuid
+import shutil
 
 __LOGGER_NAME__ = "handwriting-to-text-slack-bot"
-__VERSION__ = "0.1"
+__VERSION__ = "1.0"
+__SAVE_DIR__ = "/save/"
 
 def main():
     global logger
@@ -31,6 +33,11 @@ def main():
     if not os.environ.get('AZURE_COGNITIVE_SERVICES_URL'):
         sys.exit("No environment variable \"AZURE_COGNITIVE_SERVICES_URL\" found.")
     AZURE_COGNITIVE_SERVICES_URL = os.environ.get('AZURE_COGNITIVE_SERVICES_URL')
+
+    if not os.environ.get('SAVE_TO_DISK'):
+        SAVE_TO_DISK = False
+    else:
+        SAVE_TO_DISK = os.environ.get('SAVE_TO_DISK')
     
     bot = Slackbot(SLACK_BOT_TOKEN, SLACK_BOT_NAME)
 
@@ -40,27 +47,35 @@ def main():
     if bot.slack_client.rtm_connect():
         logger.info("StarterBot connected and running!")
         while True:
-            command, channel, user, is_AT_bot, url_private_download, permalink = bot.parse_slack_output(bot.slack_client.rtm_read())
+            command, channel, user, is_AT_bot, url_private_download, permalink, filename = bot.parse_slack_output(bot.slack_client.rtm_read())
             if command and channel and user:
-                if url_private_download and permalink:
-                    #A file was uploaded
-                    logger.info("A file was posted")
-                    filename = uuid.uuid4()
-                    logger.info("Generating filename: {}".format(filename))
-                    logger.info("Downloading from {} into local file {}".format(url_private_download, filename))
-                    bot.downloadFile(filename, url_private_download)
-                    logger.info("Getting Annotations from Microsoft Cognitive Services")
-                    result = projectOxfordHandler.getResultForImage(filename)
-                    text_filename = projectOxfordHandler.getTextFileFromResult(result)
-                    annotated_image_filename = projectOxfordHandler.showResultOnImage(result, filename)
-                    logger.info("Sending back Annotated image")
-                    bot.send_file(channel, annotated_image_filename)
-                    logger.info("Sending back Text File")
-                    bot.send_file(channel, text_filename)
-                    logger.info("Deleting temp files {} and {} and {}".format(filename, annotated_image_filename, text_filename))
-                    os.remove(filename)
-                    os.remove(annotated_image_filename)
-                    os.remove(text_filename)
+                if url_private_download and permalink and filename:
+                    try:
+                        #A file was uploaded
+                        logger.info("A file was posted")
+                        logger.info("Generating filename: {}".format(filename))
+                        logger.info("Downloading from {} into local file {}".format(url_private_download, filename))
+                        bot.downloadFile(filename, url_private_download)
+                        logger.info("Getting Annotations from Microsoft Cognitive Services")
+                        result = projectOxfordHandler.getResultForImage(filename)
+                        text_filename = projectOxfordHandler.getTextFileFromResult(result, filename)
+                        annotated_image_filename = projectOxfordHandler.showResultOnImage(result, filename)
+                        logger.info("Sending back Annotated image")
+                        bot.send_file(channel, annotated_image_filename)
+                        logger.info("Sending back Text File")
+                        bot.send_file(channel, text_filename)
+                        if SAVE_TO_DISK:
+                            logger.info("Moving files to storage")
+                            shutil.move(filename, __SAVE_DIR__ + filename)
+                            shutil.move(annotated_image_filename, __SAVE_DIR__ + annotated_image_filename)
+                            shutil.move(text_filename, __SAVE_DIR__ + text_filename)
+                        logger.info("Deleting temp files {} and {} and {}".format(filename, annotated_image_filename, text_filename))
+                        os.remove(filename)
+                        os.remove(annotated_image_filename)
+                        os.remove(text_filename)
+                    except Exception:
+                        logging.exception("Something went wrong")
+                        bot.send_message(channel, "Something went wrong")
             time.sleep(READ_WEBSOCKET_DELAY)
 
 def setup_logger():
